make it idempotent

+++ vim/install.sh
@@ -11,7 +11,7 @@ then
   [ -d ~/.vim/after ] || mkdir ~/.vim/after
   [ -d ~/.vim/after/plugin ] || mkdir ~/.vim/after/plugin
   cp vim/tabular.vim ~/.vim/after/plugin
-  ln -s "${BASE}/vim/vundles.vim" ~/.vim/vundles.vim
+  [ -f ~/.vim/vundles.vim ] || ln -s "${BASE}/vim/vundles.vim" ~/.vim/vundles.vim
   vim -u ~/.vim/vundles.vim --noplugin -N "+set hidden" "+syntax on" +BundleClean +BundleInstall +qall
 else
   echo "You should probably have installed Vim before this."
